+++ ui/spotlight-sandbox/eslint.config.js
@@ -1,0 +1,9 @@
+module.exports = [
+	{
+		ignores: [
+			'node_modules/*',
+			'build/*',
+			'dist/*'
+		]
+	}
+];